node: kubetest2: clear usage of skip-regex
https://github.com/kubernetes/test-infra/pull/34563 reminded
us about the skip regex defaults of the `kubetest2` `node` tester:

Turns out that the `kubetest2` node tester, which we need to use
in these lanes, has a `SkipRegex` default setting which include the
`Serial` tag.

Problem: all the e2e_node suites are serial, because they need to mutate
the node, which is a giant shared piece of state.

As followup, we take care to initialize the `skip-regex` option
to empty value and we demand to the label filter which test to run,
avoiding sources of conflicts.

Signed-off-by: Francesco Romani <fromani@redhat.com>