DRA: allow slow tests in E2E node CI jobs
We want to skip slow jobs only in the pre-submits and must run them in the
periodic jobs. This was wrong for E2E node jobs, but luckily made no difference
because we don't have any such test marked as slow.

The "label_filter" variable was only used in one place. The template is more
readable when directly specifying the label filter. This also enables
adding the "&& Slow" conditionally.