integration: add variants with race detection
Running with race detection is possible in our CI from a performance
perspective and useful because it tends to reveal problems that don't show up
in the simpler unit tests. We don't have race detection enabled for E2E tests.

The plan is to treat these additional jobs as an experiment until the situation
is more stable. Right now there are known problems and flakes. Once stable,
we can either enable race detection in the normal integration testing or
migrate the race variant to release informing/blocking.