DRA: simplify Jinja template
That "file" is available wasn't documented and not immediately obvious because
the comment called it "<suffix>". A plain boolean is simpler and works as long
as the suffix is doesn't conflict with other settings, which can be assured
when picking it.

While at it, use some more line breaks in the dra.jinja template.