replace loadbalancer blocking job
cloud-provider-kind was created to improve the loadbalancer logic
in the code base and to avoid the dependencies on specific
implementations and remove the flakiness introduced by the
infrastructured.

In addition, the ingress-gce job is unnmaintained, the cloud provider
specific code was not updated since the last year and the
infrastructure flakes are not debuggable.