kind jobs: sync presubmits and periodics
The naming is now consistently [ci|pull]-kubernetes-e2e-kind[-variant],
with the variant part empty for the base job. In other words:
- ci-kubernetes-kind-e2e-parallel -> ci-kubernetes-e2e-kind
- ci-kubernetes-kind-ipv6-e2e-parallel -> ci-kubernetes-e2e-kind-ipv6

The periodic jobs still used FOCUS/SKIP. Now they run exactly the same
tests as the presubmits. The risk of new flakes in the periodic jobs
is slow because the same test already were tested regularly in the
presubmit. Besides, if there are such flakes we want them to show up
when triaging the periodic jobs.

Note that slow tests continue to be disabled in the periodic jobs. This might
be worth revisiting later.