DRA canary: experiment with kubelet version skew
The intent is to run the current set of tests against a cluster
where the kubelet is from a previous release. Which tests will
pass remains to be seen...

While at it, avoid repeatedly dumping intermediate kind_yaml variable
assignment into the job's log output. Instead, manipulate the
content once and dump the final result.