DRA canary: switch from mixed kind images to kubelet replacing
The approach with using different node images didn't work: kind ended
up deploying kube-proxy with the image version which was only available
in the control-plane node, so kube-proxy didn't come up on the worker
nodes.

The approach used elsewhere is to bring up a uniform kind cluster
and then hack it by replacing binaries.