DRA canary: filter by kubelet version, use stable release for presubmit
As expected some tests fail because they test aspects of the kubelet
which were added in more recent releases, like seamless upgrades.
That test can labeled with KubeletMinVersion:1.33 to avoid running it with
kubelet 1.32.

The CI artifacts worked okay in previous canary tests, but want we want
for presubmits are the stable releases to avoid breaking PR testing for
reasons that are not under control of the PR author.