chore: increase concurrency and reduce timeout for CAPA blocking
Increase the job concurrecy for the CAPA e2e blocking job. This is so we
don't get too much of a backlog on PRs waiting for this job.
As the job currently takes less than 1 hour we can set a timeout lower.
Also, removing the run_if_changed for the time being to make the job
unconditional.

Signed-off-by: Richard Case <richard.case@outlook.com>