chore: rollback requiring CAPA e2e blocking
Revert requiring that the CAPA blocking e2e job on presubmit. We are
seeing lots of failures because there are only 10 aws accounts for
boskos to handout and they are used by CAPA and kops. An account maybe
tied for hours with the full e2e tests and then we try and run a small
subset of tests on PR and this causes boskos resource reties to be
exceeded.

Signed-off-by: Richard Case <richard.case@outlook.com>