DRA: enable kubelet n-1 and n-2 testing
We explicitly support combining an older kubelet as far back as 1.32 with
current master (aka 1.34), as long as the beta API is enabled. To test this,
these new jobs bring up a kind cluster as usual, then replace the kubelet
binary on worker nodes after downloading it from CI artifacts (periodic jobs)
or from the latest patch release (presubmit).

This distinction ensures that presubmits don't get broken by unrelated changes
on the release branches (unlikely, but could happen) and that those changes
show up as regressions in the CI.

The new jobs are not release informing yet because they are new. They could
get promoted later.