etcd: run govulncheck using Go 1.23
Keep the Go version in sync with the Go version used by the stable
releases. Once the stable branches use Go 1.24, we can use the tool
directive and the installed tools should use the same toolchain version
from the go.mod. Therefore, we can go back to use the kubekins image,
and consolidate the jobs.