DRA: add pull-kubernetes-kind-dra-all-slow
We've had problems with PRs being merged which then broke periodic testing
because some slow test wasn't working anymore or (right now) the load became
too much.

We still don't want to include slow tests in presubmits, but having a job to
try out the full set of tests on demand is useful. Because this is a new job
and a simple variant of the existing job, both the normal and canary presubmit
get added at the same time,

While at it, the checking of all boolean config options gets
simplified. There's no need to compare against "true".