Dont use discovery store on terraform jobs
The terraform that kops generates can't handle the state store s3 bucket and discovery store s3 bucket being in different regions.

Since it is reasonable for cluster administrators to have both buckets in the same region and only prow's shared buckets are in different regions, we can just skip the discovery store in our terraform tests.