SIG testing: alpha/beta "enabled" and "conformance" presubmits
We've had kind-alpha-beta-features and kind-beta-features jobs for a while. The
original purpose was to run stable tests in a cluster with features enabled to
detect when enabling those breaks stable functionality. Later the jobs were
extended to also run all tests which should work in such a cluster. The
kind-alpha-features job got removed recently because it's not necessarily a
valid cluster configuration.

What this adds for the both cluster configs is:
- "-enabled": running only tests for on-by-default features, i.e. excluding tests for
  stable features. This matches the original purpose of the jobs.
- "-enabled-conformance": restricts the test selection even further to only
  conformance tests.

Both can eventually get promoted to release informing or even blocking.

Only presubmits get added for now. If testing the jobs in a trial PR works, the
corresponding periodics can be added.