SIG testing: clean up alpha/beta presubmits
It should no longer be necessary to skip specific tests. If they fail, then
they are either broken and we want to know, or they are not labeled correctly
and we want to fix that.

This intentionally only changes the presubmits for a trial run in a test PR.
If that works, the periodics need to be updated accordingly.