SIG testing: use only label filter for alpha/beta jobs
The implicit default for SKIP is '\[Serial\]'. That's surprising ("what you see
is *not* what you get") and unnecessary for these jobs. If we have serial tests,
then we can and should run them to get full coverage.

There is a risk that too many serial tests slow down the overall job. If that
is the case, then skipping serial tests should be added back inside the label
filter.

FOCUS doesn't have a default in these jobs, but elsewhere it defaults to
'[Conformance]'. It's safer to explicitly set it - perhaps someone does a
copy-and-paste of these settings.