Make node E2E presubmit optional for containerd
K8s node E2E tests now run via Github Actions on containerd pull
requests, as part of the containerd migration off of Prow:
https://github.com/containerd/containerd/issues/11486

Given that these tests are stable, we can proceed with marking these
Prow tests as optional on containerd pull requests.

The job for the 1.6 branch was left untouched because the node E2E tests
that run as Github Actions do not work on that branch, however 1.6 is
past EOL and so we can remove this job once we remove the other
containerd jobs.

Signed-off-by: Chris Henzie <chrishenzie@google.com>