kind-alpha-beta-features: not release informing, serial experiment
The experiment is meant to determine whether running serial tests in the same
job as the parallel tests works and how much slower it becomes.

Even long term the job probably shouldn't be release informing because the
alpha tests can break it at any time, without affecting the release. What needs
to become release informing instead are the kind-alpha-beta-enabled and/or
kind-alpha-beta-enabled-conformance jobs (to be determined after we are done
with setting everything up and they proof to be stable).