sig-testing: remove KUBE_GORUNNER_IMAGE override
To support running dynamically linked control plane components,
the base image had to be
overridden. https://github.com/kubernetes/kubernetes/pull/133972 ensures that
this happens automatically, so we no longer need to do it in the job. That's
good because it's not clear in the job which image should be used.