SIG testing: disable serial tests in canary
Ginkgo now has support for scheduling slow tests sooner in a parallel
run. To see the effect of that we need a CI job which includes slow
tests. The canary job is right for that, but including serial tests affects the
overall runtime too much at the moment. Let's disable them, experiment with
prioritization of slow jobs, then come back to serial later.