DRA: better support for kind.yaml changes
A recent change to test/e2e/dra/kind.yaml did not start the DRA jobs, so it
didn't get noticed that the change broke the jobs. We need to include YAML
in the set of files which may trigger jobs.

The kind.yaml change was reverted for now, but will come eventually.  The
canary jobs get updated to support that change without breaking current master.