Add insecure registry configuration for localhost:5000 in CRI-O e2e tests
This configures CRI-O to allow HTTP connections to localhost:5000,
which is required for e2e tests that use the fake-registry-server.
The fake-registry-server only serves HTTP, but CRI-O defaults to
HTTPS for all registries, causing image pull failures.

The configuration is placed in /etc/containers/registries.conf.d/
as a drop-in file to avoid being overwritten by the CRI-O
installation process.

This fixes test failures in:
- [sig-node] Ensure Credential Pulled Images
- [sig-node] Container Runtime Conformance Test (private registry tests)

Related: kubernetes/kubernetes#134998
Related: kubernetes/kubernetes#135369
Signed-off-by: Sascha Grunert <sgrunert@redhat.com>