alpha/beta race detection: enable explicitly
The indirect enablement through a sentinel test didn't work because it's not
possible to determine before a run which tests are enabled. Now an explicit
parameter is used.

Support for KUBE_E2E_TEST_ARGS in Kubernetes still needs to be reviewed and
merged. This job update is meant to verify that it works.