SIG testing: define test-infra job for local-up.sh
https://github.com/kubernetes/test-infra/pull/36028 broke
https://testgrid.k8s.io/conformance-all#local-up-cluster,%20master%20(dev) and
https://testgrid.k8s.io/sig-node-dynamic-resource-allocation#ci-dra-integration (both
using local-up-cluster.sh).

Instead of merging an image bump blindly and hoping that it goes well, let's do
at least some trial runs with jobs that will be affected by an image
bump. The new pull-test-infra-local-e2e is such a job. It gets triggered by
edits to the job file (like image bumps) and is optional (can be ignored if the
normal job is unstable).